refactor(ui): extract WorkflowDetailView mirroring ApiDetailView
`WorkflowDetailPage.tsx` was an 818-line page that mixed seven
concerns: route plumbing (params, navigation, scroll restore,
escape-to-back), the workflow query and delete mutation, the page
header, the imported-workflow body (meta strip, slug chip, tab
toggle, lazy Arazzo viewer with error boundary), and a 100+ line
catalog-import fallback. Split it the same way `ApiDetailView` was
just split, so both detail surfaces now share an identical shape:

   pages/<X>DetailPage.tsx → orchestrates routing + queries + header
                              + delete dialog
   components/<area>/<x>-detail/<X>DetailView.tsx → renders body
   components/<area>/<x>-detail/<peers>.tsx       → leaf sections

The page drops from 818 → ~190 lines and reads like a recipe of
what's on screen. The view file owns the URL-backed tab state
(`?view=…`) — the only piece of state that lives below page level —
so swapping between Overview / Diagram / Docs / Split is now
contained in one file alongside the lazy `ArazzoUI` import.

No behaviour change. All 443 unit tests + 33 e2e specs still pass.

- New folder `components/workflows/workflow-detail/`:
  - `WorkflowDetailView.tsx` — meta strip + slug chip + tab toggle
    + Arazzo viewer (lazy) for imported workflows
  - `OverviewBody.tsx` — two-column overview pane (steps + APIs)
  - `StepRow.tsx` + `StepRowBody.tsx` — single Arazzo-step renderer
  - `SlugChip.tsx` — copy-on-click slug badge
  - `ArazzoErrorBoundary.tsx` — graceful failure for the viewer
  - `CatalogWorkflowFallback.tsx` — un-imported catalog workflow
    surface, with an optional `helpSlot` prop so the page can pass
    its `<PageHelp>` into the header
  - `index.ts` — re-exports `WorkflowDetailView` and the fallback
- Slim `pages/WorkflowDetailPage.tsx` to delegate the body to the
  new view component while keeping all routing / mutation / help
  plumbing at the page level
