+++ ui/src/components/workflows/workflow-detail/index.ts
@@ -1,0 +1,3 @@
+export { CatalogWorkflowFallback } from './CatalogWorkflowFallback';
+export { WorkflowDetailView } from './WorkflowDetailView';
+export type { DetailTab } from './WorkflowDetailView';
